RESTful naming update
Updated the names of the endpoints that control the creation of nodes to better align with the RESTful naming conventions. The structure will be as follows, the network resource is responsible for creating and broadcasting to the entire network and the node resource is responsible for creating stuff at the given node.