ads to  processAndSend list
lists added on node level for shipments that are currently at the node and need to be sent on to another node and a list for deliveries that have been finalized at the current node.

when a shipment is created and broadcasted to the network it is added to the send list at the current node.

there needs to be a lot of error handling but that's not a part of the assignment but might be added later